Fix for enter key submit issue and filter fields spacing
-Added timeout to resolve enter key problem: The issue was the start/end
date validation on whether the input data was valid or not slowed down
angular's triggering a digest cycle, it was trivial to 'beat the clock'
and hit enter to search before the date field appropriately reflected in
the model. A small timeout provides imperceptible wait time for
insurance against the model being updated.

-Adjusted sizes and spacing of elements on flight and passenger filter
inputs. Alignment checks out in Chrome/FF on local testing.