Fix for issue #87
-applied fix for date-time range issue.

-UTC proper conversion takes place just before front-end service calls
are made now. The conversion changes the start time to begin at 0:00:00
and end-time to be at 23:59:59. In order to not affect the front-end
visually a temporary object is created that represents the model with
the altered time. This was due to setting a date to UTC can sometimes
push a date backward or forward depending on the GMT offset, this would
skew the front-end look even if it programmatically did the correct
action