Tickets #352, #373, #312, #275, #344, #287, #236
-added iterator over documents in passengerServiceImpl in order to grab
document information for the passenger list for #275

-Added edifactmessage to target in pnrserviceimpl.java return in order
to get transmission date for paxdetail page #344

-Adjusted passengerdetailscontroller.java to use the newly available
edifactmessage to grab the transmission date out to pass through the
passengerVo.

-added new properties for naming

-added new functions to PaxController.js in order to in sure order for
TVL lines based on etd #312 and to remove flights from flight history
that were not present in the current pnr #373

-adjusted pax.detail.html to display new Transmission Date with time
stamp #344, to display street address #352, and some small text
adjustment to  reduce confusion for now on Flight History as to what we
consider the itinerary coming from #236

-Also, changed pax.detail.html to use the etdLocalTZ element in order to
keep consistency between the PNR tab and the flight history tab. This
should be revisited in the future in order to insure we're showing the
times we actually intend to.