Part 2 of 3. Prime flight idea incorporated, no booking detail save yet
-LoaderMessageReceiver modified to use application context in order to
get a new instance of LoaderWorkerThread per use instead of it being a
singleton managed bean

-returned LoaderRepo flight processing method to form, as well as
removed synchronization due to prime flight incorporation

-a "placeholder" string has been added to the base loader functions in
order to manually allow the loader to run. WARNING: Manually running the
loader will revert to the original idea of loading all flights on a pnr
as flights as the listener manages the prime flight indicator which is
bypassed by manual load.