Fixed APIS loading, linked pnr and booking details
-added findOne() lookup for PNR inside of loader repo. The pnr was
detached from the transaction the booking details are created in. It is
bidirectional, so PNR had to associate to the details and vice versa,
while also maintaining the same for passengers. In order to get an
attached pnr in order to associate to booking details, the pnr is
fetched live from the DB. **This may not be a final solution, as
expanding the transactional state might suffice for several reasons.

-added malformed prime flight key check to loader utils. This will
enable badly formed prime flight keys to just dump their message payload
into the prime flight tables for now.

-removed old, now erroneous pnr/booking association in pnr message
service

-Adjusted integration test to pass -1 for the required PNR id. The same
was added to ApisMessageService calls since there is no PNR id for those
messages obviously.