Testing expanded transacational state
Current theory: Transactional states scopes in loader repo are causing
detached entities to be associated to the message to be saved, so that
when it is time to save hibernate pulls down those entities once again
in order to do so.

Potential solution: Expanding transaction scope to insure, throughout
message process, the state never becomes detached for any of the
objects.

Local machine has seen 33%~ increases in load time with this.

-Putting transactional state at higher level in order to restrict the
possiblity
-commenting out existing sub-states in loader repo