Fix to querybuilder query results
Issue: New non-hibernate managed mapping caused issues as old JPQL
expected to be able to use "join <entity>.<mapped-entity>". Such as
flight.passengers. This is not possible, and as such the particular
table joins that are needed for specific queries (i.e. get flights where
passenger documents = docNumber) needed to be re-written to self-joins
and/or cross joins.

-Query builder operates on being able to query terms for flights or
passengers and as such JPQL generator had both sides of the programmatic
construction of the queries altered to now use the aforementioned
cross-joins/self-joins

-Constants had "exists" added as a keyword constant for query creation
use