Loaderworkerthread poll time changed for APIS messages
Pure APIS message dumps take much longer under new threading
methodologies due to the time a queue is kept alive in order to check
for incoming messages (to accomodate for bursts of messages, we would
not want to destroy queues only for a few seconds later to have to
re-create the same queues because of a short delay in same prime-flight
messages delivery)

APIS usually only comes in one message per prime-flight however, making
it wait the entire poll time for a single message that created a new
queue. This is prohibitively long for mostly APIS dumps.

Instead of waiting 30000ms, the thread will now only poll for 5000ms
before consuming the queue and enabling another one to be processed.