Tamr integration
-This is the initial messaging system integration between GTAS and TAMR.
Currently all addresses are empty and connections are commented out.
With appropriate connection credentials and uncommenting this can be
reinstated. This is a completely and totally optional integration and
its default state is unused.

TamrConversionService/...Impl
-Service for converting GTAS database items into Tamr specific POJOs in
order to be converted to the appropriate JSON message.

TamrMessageReceiver
-JMS Listener, listening to outbound queue from tamr and depositing
results into our tables

TamrMessageSender
-JMS Sender, sending to inbound queue for tamr

TamrQueueConfig
-Configuration file for queues

TamrDocumentSendObject/TamrPassengerSendObject
-Passenger/Document pojos for sending converted JSON objects.

TamrResponse/TamrResponseData
-Receiving pojos to enable reading of incoming Tamr messages