This commit fixes the issue in #901.
Matching Service : Got rid of un-needed public access modifier.

Matching Service Impl : We were setting watch list item name and dob based on the passenger that was a hit instead of the watchlist item. This resulted in the error seen on the front end. Changed to these values from watchlist item