Issue #924
There was a check on the destination equaling the origin of the flight which
messed up the dwell time for round trip tickets.

a -> b -> a would not record the dwell time of b -> a.

We have decided to loosen the application to record the dwell time of b -> a.