Ticket 801 - added a regex check on segments to insure they start in … (#949)
Issue #801 
Updated Edifact Lexer to check to see if a segment starts with 3 letters and a terminator OR separator token.

Updated PnrUtils to put a NF for a not found gender. Bug fix on how gender was found. Also, will take last available gender when iterating over ssr docs (there was an issue if the travel doc was F or FI).