AWS SQS integration, resolves #951
A scheduled task to process messages from AWS SQS queue.

The initial delays in milliseconds and the fixed period between
invocations can be configured in the jobScheduler.properties file,
as well as the flag to disable/enable it.

The scheduler will read messages from AWS SQS then pushes
the messages to REDIS to start the loading process.