Updated fuzzy matching. Performance increase from 20 seconds on a run to ~5 seconds with 521 passengers.
Added a class to pass in parameters.

Moved all database calls related to running fuzzy matching on a passenger out of a for loop.
There will still be a call *inside* fuzzy matching to create a case and watchlist link if there is a hit.

Updated sql to pull back passengers on flight instead of a service call. Used join fetch to hydrate data.