#1258 Parser Tuning on ADD phone numbers.
ADD was not to specification. The specification says we should parse phone numbers from free text.

It is suspected ADD was written to parse PNR 11.1 but we do not have a spec doc.

Updated the ADD to parse for (**suspected**)11.1 and 13.1 PNR.

Added unit testing. Made it clear that the numbers we are parsing are both fake AND obviously phone numbers.