move svg path code to service (??), make graph links more clickable, misc refactoring
vertical links between nodes are created without any background fill (which would be clickable), so there's only a 1 pixel thick target to close a link when there's no text. Need to add transparent clickable background to links, but for now upped the stroke width to 4. Also styled links on hover back to red with a cross.