#1347 flight page countdown timer.
Encapsulated countdown logic that was in case class into it's own class called count down calculator. Modified both casevo and flightvo classes to use the same class/logic.

Added a view to retrieve the relevant ETA or ETD of a flight. Added rollback and update to view.

Added some hibernate properties and schema filters so when creating or dropping the database it would ignore the view.

Added a vo for the countdown object.

Added countdown time to the ui-grid.

Added relationship between flight and view.