#1426 Update the passenger repository to generate a sql statement for the front end.
Add 2 indexes to increase performance.

Added a "hit view" which is similar to a case in that it holds status of rule results. It is bound to hit details instead of passengers.

Changed name of case page to PriorityVettingList to better explain what the purpose is for.

Updated rule engine to automatically create hit views on hit details per each group.

Added default group to gtas_data.sql

Modified cases to work with date times instead of just times.