End to end functionality now works.
Add email to user to have users toggle in to receive automated notifications.

Rollback TestCommonServicesConfig changes as they were overriding values outside of test.

Change duration formatting to use count down calculator. Duration formatting broke when using a historical date.

Added class in passenger repository - this will bring the passengers to a managed state in hibernate without updating the entity.

Added logging when sending a automated email message - this will be removed later but we need to track it, if only in logs for now.