Partially completed login page for GTAS.
Updated serviceWrapper.js and genericService.js to make a post form authentication request.

Updated the form to have a callback to propagate data to the login page. Updated login page to redirect to about gtas when completed.

Ability to customize urls, and more importantly a bug in the login page that is currently suppressed by "no throw" needs to be handled and the navigation logic needs to be updated to have the history of the login page.

Updated packages to include prop types as well as js-cookie.