Allow filterForm to accept a method that will customize the fetch it makes allowing for users to control how filter form fetches data.
Added a few tweaks to the table that will likely be blown away later to get data displaying for flights - the ability to ignore a field as well as switching the data to be parsed every update, and only updating if the data changes.

Updated backend to accept a get request instead of a post.

Tweaked flights to allow for values to populate correctly for call to GTAS.

Data should now flow to the flights table from GTAS.

Known issues -> back end pagination is not supported and so the first 10 records is always hard-coded.