Made ETL Updates:
1. Changed Transformations to read from their own temp file, instead of
reading the common temp file (which saves a lot of reading, sorting, and
writing time)
2. Adjusted sorting threshold at 35% for different steps of the
transformation (Sorting will be done in memory for a given step of a
transformation till the available memory goes down to 35%). If available
memory is below 35%, sorting for the step on which processing is being
done will be moved to the disc till the memory becomes free.
3. Added processing summary at the end of the log.
4. Changed Transformations to execute in parallel for 100% mutually
exclusive operations