The FlightPax class combines several different tables into a single class. This confused where the actual data was and so it was decided to remove the table.
Due to the structure having links to both flight and passenger, apis information and baginformation this class was used in several places throughout the application.

The work encompassed here deprecates the FlightPax and removes it from business logic in the application.

I kept the class in the rule engine as well as kept the logic to load into the database. This will be removed in a future release.