Backend update to allow for sending of a passenger.
This will allow the gui to request that the rule engine send a passenger, note, and latest relevant message to another system utilizing the gtas-information-share pojos. This will create a new external rule hit in the participating system.

This also incorporates new parsing to parse in passengers who contain notes.