Some of the information GTAS ingest from PNR data consist of free text fields.
It appears that these fields could follow some degree of structure, but this is not defined or enforced in any meaningful way. Because of this, there is no way to reliably extract contact information or other informative pieces.

To allow for more of a manual search of this certain segments, the parser has been enhanced to take a copy of none to all of the segments IFT,SAC,LTS,SSR,EBD, and ADD. To allow selectivity when selecting segments they must match a list of provided regex. Both the segments and the provided regex are defined in the default.application.properties. Note that only the first matching regex will be recorded by the saved segment. By default no segments are saved.

Once the segments have been extracted they are saved to the saved_segments table of GTAS. Rules and queries can be run on the raw messages, the segment name, and the regex that caused the segment to be saved.

Due to the nature of the free text the only query and rule functionality provided is "contains" and "not contains."