Update rule runner to go horizontal.
These code changes are to assist in horizontal drools rules. Instead of
using vertical scaling threads the rule engine for drools switch over to
a queue based system.

Hits are also always persisted in an async manner.

This allows the rule engine to process messages on a queue instead of a
scheduled.

Some  infrastructure changes were also done in order to support the new
rule runing application