Add horizontal loading.
Update GTAS to allow for horizontal loading, instead of just vertical.
Break out loading logic into loader workers.
Keep scheduler as the mediator and first point of contact for all
messages.
It is critical that GTAS only process a flight on a single thread across
all loaders. The scheduler is responsible for ensuring all loader
workers have access to different flights, and none are being worked at
the same time.