Convert PVL to use Hit View Status table as root
This allows for the use of indexes and shortening queries by eta/etd
date based on flight.

The query will throw an exception if it does not have eta/etd.

This should allow for most use cases of the PVL to be quickly performed.
There is still an issue with longer running queries.