De-duplicate Watchlist Items
The watchlist items all follow the same structure and can be
de-duplicated in the database by saving the components as a string.
There are two unique values

1) first name + lastname + dob + hit category
2) doc number + doc type + hit category

Once the key is saved in the database a quick check can be run against
it and duplicated values can be logged, ignored, and returned with a
duplicate status. This will relieve pressure on the drools engine, which
does not perform well when the several hundred duplicate watchlist items
are included on a knowledge base.